fix(devtools-connect): give application code for custom OIDC HTTP options precedence
Instead of using the application callback first and then potentially transforming the
result with options computed by devtools-connect, do the reverse. This gives the
application using devtools-connect more flexibility about how exactly to modify OIDC
HTTP options.

The goal here is to enable usage of HTTP proxies in conjunction with `useSystemCA`
and explicit `tlsCAFile` settings; devtools-connect correctly passes the former
to the OIDC plugin, but after this change, the application can also pass it to
the proxy agent, if it creates one.

None of our team's applications currently make use of the function variant of this
options, so this will not affect existing code in any way.