fix(signing-utils): migrate to new Windows signing cert DEVPROD-13200 (#493)
DEVPROD-13200: Migrate to new Windows signing cert

This commit updates the certificate we use to sign to our new one for
the next three years. Ideally we would also update this to be
parameterized, but given that there are a few projects that depend on
this script, I think it's safer to make this update as-is now and decide
on templating later.