fix(devtools-connect): re-try connection in case of TLS errors without system CA MONGOSH-1935
Recent user reports have made it clear that our attempts to solve
TLS errors caused by adding system certificates haven't fully been
resolved yet. To address this, we add a generic fallback that
attempts to connect a second time without the system certificate store
added.

This commit also adds TLS errors to the list of fail-fast errors,
i.e. errors which should result in a quick connection end because they
are unlikely to be resolved by the Node.js driver attempting to
re-connect repeatedly on the level of individual connections.
This should avoid situations in which timeouts make the connection
attempt take twice as long for TLS errors.