feat(redact): add a dictionary of secrets to redact MCP-29 (#569)
* feat(redact): add a dictionary of secrets to redact

With this feature we support redacting known secrets from a
dictionary. This is value for environments like the MCP Server where
we create new users (user/password) or when we can infer secrets from
CLI arguments (--user, --pasword).

* chore: add support for arbitrarily deep object redacting

* chore: refactor to reuse the iteration, instead of iterating twice

* chore: fix linter warnings

* chore: fix an issue when the constructor is not a function

We don't want to return a truthy value if the ctor is a non function
value (like an object). Plain objects either have a plain object as
a prototype or a null prototype