fix(mongodb-log-writer): fix flushing when messages are pending MONGOSH-2883 (#578)
The current `.flush()` method only flushes the target output stream.
This is broken behavior, because it can be run at a point in time
at which other writes to the log writer itself have already been
scheduled, but have not yet been processed (i.e. written to the
target stream).