feat(oidc-mock-provider,mongodb-runner): make OIDC mocks more broadly usable COMPASS-10034
Internal TSEs have requested making oidc-mock-provider available for
internal testing with OIDC. While it cannot replicate every aspect
of real-world identity providers, it is an easily spun up local
equivalent of those, and provides flexibility that those real-world
identity providers lack in terms of configurability.

This change widens the array of CLI options provided for the
oidc-mock-provider CLI, and integrates it into mongodb-runner
so that the latter can spin up a joint OIDC-IdP-and-mongod-cluster
environment on Linux, if that is desired.