feat(eslint-plugin-devtools): add rule forbidding non-null prototypes for `Record`s
`Record` types in TS are typically used to indicate that an object
may have arbitrary keys. This means that using `{}` or similar object
literals to initialize them is a poor choice -- if user code can
cause keys like `hasOwnProperty` or `__proto__` to be used as
keys on those objects, they will not behave as intended.

This has been the cause of many security issues in the JS
ecosystem, so let's add a lint rule to limit impact
of these issues and force ourselves to stick to best practices.